chore: maintain the IIFE wrapper used by rollup as an independent file
